+++ app/src/main/java/com/sysadmindoc/callshield/data/BackupRestore.kt
@@ -157,6 +157,7 @@ object BackupRestore {
         val smsContentEnabled: Boolean = true,
         val smsBurstEnabled: Boolean = true,
         val urlhausStripQueryEnabled: Boolean = true,
+        val urlhausRemoteLookupEnabled: Boolean = false,
         val contactWhitelistEnabled: Boolean = true,
         val contactsOnlyEnabled: Boolean = false,
         val dbPrefixExpansionEnabled: Boolean = false,
@@ -1117,6 +1118,7 @@ object BackupRestore {
             smsContentEnabled = this[SpamRepository.KEY_SMS_CONTENT] ?: true,
             smsBurstEnabled = this[SpamRepository.KEY_SMS_BURST] ?: true,
             urlhausStripQueryEnabled = this[SpamRepository.KEY_URLHAUS_STRIP_QUERY] ?: true,
+            urlhausRemoteLookupEnabled = this[SpamRepository.KEY_URLHAUS_REMOTE_LOOKUP] ?: false,
             contactWhitelistEnabled = this[SpamRepository.KEY_CONTACT_WHITELIST] ?: true,
             contactsOnlyEnabled = this[SpamRepository.KEY_CONTACTS_ONLY] ?: false,
             dbPrefixExpansionEnabled = this[SpamRepository.KEY_DB_PREFIX_EXPANSION] ?: false,
@@ -1220,6 +1222,7 @@ object BackupRestore {
         preferences[SpamRepository.KEY_SMS_CONTENT] = smsContentEnabled
         preferences[SpamRepository.KEY_SMS_BURST] = smsBurstEnabled
         preferences[SpamRepository.KEY_URLHAUS_STRIP_QUERY] = urlhausStripQueryEnabled
+        preferences[SpamRepository.KEY_URLHAUS_REMOTE_LOOKUP] = urlhausRemoteLookupEnabled
         preferences[SpamRepository.KEY_CONTACT_WHITELIST] = contactWhitelistEnabled
         preferences[SpamRepository.KEY_CONTACTS_ONLY] = contactsOnlyEnabled
         preferences[SpamRepository.KEY_OUTGOING_RISK_WARNING] = outgoingRiskWarningEnabled